Revert "refactor: main components를 세부 폴더로 분리"
This reverts commit fe0c07d579ce26a374a61994917fd2d26e568ae3.
