+++ src/app/[locale]/home/components/MobileHomeHeader.tsx
