+++ src/app/[locale]/home/components/PlaceList.tsx
