+++ src/app/[locale]/home/components/ThemeItem.tsx
