+++ src/app/[locale]/home/components/TravelItem.tsx
