+++ src/mocks/handlers/place.ts
@@ -1,9 +1,9 @@
 import { HttpResponse, http } from "msw";
 
-import { ResponsePlace } from "@/lib/types/place";
+import { IResponsePlace } from "@/lib/types/place";
 import { placeResponse } from "../response/place";
 
-export const getPlaces = http.get<never, never, ResponsePlace>(
+export const getPlaces = http.get<never, never, IResponsePlace>(
   `${process.env.API_HOST}/place`,
   () => {
     return HttpResponse.json(placeResponse);
